+++ app/directives/index.ts
@@ -9,7 +9,7 @@ import highlights from './highlights';
 import kramjax from './kramjax';
 import onOutside from './onOutside';
 import onTextSelect from './onTextSelect';
-import pdfPage from './pdfPage';
+import pdf from './pdf';
 import subnav from './subnav';
 import toc from './toc';
 import validateDocumentSource from './validateDocumentSource';
@@ -28,7 +28,7 @@ export default function(app) {
   kramjax(app);
   onOutside(app);
   onTextSelect(app);
-  pdfPage(app);
+  pdf(app);
   subnav(app);
   toc(app);
   validateDocumentSource(app);
