+++ package.json
@@ -81,6 +81,7 @@
       "angular-sanitize": "github:angular/bower-angular-sanitize@^1.5.0",
       "jquery": "npm:jquery@^2.2.0",
       "lodash": "npm:lodash@^4.2.0",
+      "mutx": "npm:mutx@^0.1.0",
       "ngSmoothScroll": "github:d-oliveros/ngSmoothScroll@^2.0.0",
       "paperhive-sources": "npm:paperhive-sources@^4.0.2",
       "pdfjs-dist": "github:mozilla/pdfjs-dist@^1.4.37",
